Fixed #31286 -- Defaulted to running checks against all databases.
Regression in 0b83c8cc4db95812f1e15ca19d78614e94cf38dd which added support for
running checks only against databases but also defaulted to running against no
databases if not specified.

Note that this continues to *not* default to runing database tagged checks for
all management commands except the migrate one as whether or not we should
change this must be discussed further.

Thanks Tim Graham for surfacing how this was a problematic default.